add Mat.fromString methods, create Mat from toString data
