+++ tests/shared/src/test/scala/MatExtensionsTest.scala
@@ -16,7 +16,7 @@
 
 import slash.vector.Vec
 import slash.matrix.*
-import slash.matrix.Mat.*
+//import slash.matrix.Mat.*
 
 class MatExtensionsTest extends munit.FunSuite {
   test("Mat[1, N] -> Vec[N] -> Mat[1, N]") {
